Apply clang-tidy fixes

+++ src/benchmarks/split_complex.cpp
@@ -8,6 +8,7 @@
 #include <algorithm>
 #include <chrono>
 #include <cmath>
+#include <cstddef>
 #include <cstdio>
 #include <numeric>
 #include <string_view>
@@ -39,8 +40,8 @@ auto timeit(std::string_view name, size_t sizeOfT, size_t N, Func func)
         all_runs[i] = std::chrono::duration_cast<microseconds>(stop - start).count();
     }
 
-    auto const runs            = std::span{all_runs}.subspan(margin, all_runs.size() - margin * 2);
-    auto const avg             = std::reduce(runs.begin(), runs.end(), 0.0) / double(runs.size());
+    auto const runs = std::span{all_runs}.subspan(margin, all_runs.size() - static_cast<std::size_t>(margin) * 2);
+    auto const avg  = std::reduce(runs.begin(), runs.end(), 0.0) / double(runs.size());
     auto const itemsPerSec     = static_cast<int>(std::lround(double(size) / avg));
     auto const megaBytesPerSec = std::round(double(size * sizeOfT) / avg) / 1000.0;
     std::printf("%-32s avg: %.1fus - GB/sec: %.2f - N/usec: %d\n", name.data(), avg, megaBytesPerSec, itemsPerSec);
